backend_challenge/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pip==22.3
PyJWT==2.8.0
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2024.1